etl/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "spicy-regs-etl"
version = "0.1.0"
description = "ETL pipeline to convert Mirrulations JSON to Parquet on R2"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "polars>=1.0.0",
    "tqdm>=4.0.0",
    "httpx>=0.27.0",
    "sentence-transformers>=5.2.0",
]

[project.optional-dependencies]
embed = [
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
]

[project.scripts]
etl = "etl:main"
embed = "embed:main"
spicy-regs = "cli:main"

[tool.setuptools]
py-modules = ["etl", "upload_r2", "download_r2", "embed", "cli"]

